feat(trace-view): Add ai spans tab
