+++ static/app/views/insights/agentMonitoring/utils/referrers.tsx
@@ -2,4 +2,5 @@ export enum Referrer {
   MODELS_TABLE = 'api.performance.agent-monitoring.models-table',
   TOOLS_TABLE = 'api.performance.agent-monitoring.tools-table',
   TRACE_DRAWER = 'api.performance.agent-monitoring.trace-drawer',
+  TRACES_TABLE = 'api.performance.agent-monitoring.traces-table',
 }
