+++ static/app/views/insights/agentMonitoring/utils/types.tsx
@@ -1,0 +1,6 @@
+import type {TraceTree} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
+import type {TraceTreeNode} from 'sentry/views/performance/newTraceDetails/traceModels/traceTreeNode';
+
+export type AITraceSpanNode = TraceTreeNode<
+  TraceTree.Transaction | TraceTree.EAPSpan | TraceTree.Span
+>;
