+++ static/app/views/performance/newTraceDetails/useTraceContextSections.tsx
@@ -3,9 +3,11 @@ import {useMemo} from 'react';
 import {VITAL_DETAILS} from 'sentry/utils/performance/vitals/constants';
 import useOrganization from 'sentry/utils/useOrganization';
 import type {OurLogsResponseItem} from 'sentry/views/explore/logs/types';
+import {hasAgentInsightsFeature} from 'sentry/views/insights/agentMonitoring/utils/features';
+import {getIsAiNode} from 'sentry/views/insights/agentMonitoring/utils/highlightedSpanAttributes';
 import type {TraceRootEventQueryResults} from 'sentry/views/performance/newTraceDetails/traceApi/useTraceRootEvent';
 import {isTraceItemDetailsResponse} from 'sentry/views/performance/newTraceDetails/traceApi/utils';
-import type {TraceTree} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
+import {TraceTree} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
 
 export function useTraceContextSections({
   tree,
@@ -35,6 +37,8 @@ export function useTraceContextSections({
   );
 
   const hasSummary: boolean = organization.features.includes('single-trace-summary');
+  const hasAiSpans: boolean =
+    hasAgentInsightsFeature(organization) && !!TraceTree.Find(tree.root, getIsAiNode);
 
   return useMemo(
     () => ({
@@ -44,7 +48,8 @@ export function useTraceContextSections({
       hasTags,
       hasVitals,
       hasSummary,
+      hasAiSpans,
     }),
-    [hasProfiles, hasOnlyLogs, hasLogs, hasTags, hasVitals, hasSummary]
+    [hasProfiles, hasOnlyLogs, hasLogs, hasTags, hasVitals, hasSummary, hasAiSpans]
   );
 }
